slides/requirements.txt
# This is for netlify
PyYAML
jinja2
